pyproject.toml
[project]
name = "voiceiq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "boto3>=1.38.18",
    "fastapi>=0.115.12",
    "pika>=1.3.2",
    "pydantic-ai>=0.2.4",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "supabase>=2.15.1",
]
